% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-package.R
\docType{package}
\name{aemet-package}
\alias{aemet-package}
\alias{aemet}
\title{R interface to AEMET API}
\description{
Download climatic and meteorological data from AEMET directly from R.
}
\keyword{package}
