% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_climatology_station.R
\name{aemet_climatology_station}
\alias{aemet_climatology_station}
\title{Get normal climatology values for a station}
\usage{
aemet_climatology_station(station, apikey)
}
\arguments{
\item{station}{Station identifier code (see \code{\link{aemet_stations}}).}

\item{apikey}{Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/inicio}).}
}
\value{
a data.frame
}
\description{
Standard climatology from 1981 to 2010.
}
\examples{
\dontrun{
aemet_climatology_station("5911A")
}
}
