% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_stations.R
\name{aemet_stations}
\alias{aemet_stations}
\title{Get AEMET stations}
\usage{
aemet_stations(apikey)
}
\arguments{
\item{apikey}{String with the AEMET API key (see https://opendata.aemet.es/).}
}
\value{
A SpatialPointsDataFrame.
}
\description{
Get AEMET stations
}
\examples{
\dontrun{
stations <- aemet_stations(apikey)  #SpatialPointsDataFrame
stationdata <- as.data.frame(stations)
}
}
