% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms2decdegrees.R
\name{dms2decdegrees}
\alias{dms2decdegrees}
\title{Converts degrees, minutes and seconds to decimal degrees}
\usage{
dms2decdegrees(input)
}
\arguments{
\item{input}{character}
}
\value{
num
}
\description{
Converts degrees, minutes and seconds to decimal degrees
}
\examples{
dms2decdegrees("055245W")

}
