% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from AEMET}
\usage{
get_data(apidest, apikey)
}
\arguments{
\item{apidest}{character Destination URL. See \url{https://opendata.aemet.es/dist/index.html}.}

\item{apikey}{Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/inicio}).}
}
\value{
a data.frame
}
\description{
Get data from AEMET
}
